move render_doc_entry
